Fixed issue cause by merge conflict and readme
I fixed a missing curly bracket in the rest api stack file after
editting the file in the merge editor. I also updated the readme to
include mention of the translate endpoint now working. I also chnaged
the sort key for the movie reviews table as it was interefering with
translation working.

Signed-off-by: Saoirse O'Donovan <saoirseodonovan13@gmail.com>